+++ UI/UserControls/ProfileControl.axaml.cs
@@ -178,6 +178,7 @@ namespace PinayPalBackupManager.UI.UserControls
                     WindowStartupLocation = WindowStartupLocation.CenterOwner,
                     CanResize = false,
                     ShowInTaskbar = false,
+                    Topmost = true,
                     Background = Avalonia.Media.Brushes.Transparent
                 };
 
@@ -222,6 +223,7 @@ namespace PinayPalBackupManager.UI.UserControls
                     WindowStartupLocation = WindowStartupLocation.CenterOwner,
                     CanResize = false,
                     ShowInTaskbar = false,
+                    Topmost = true,
                     Background = Avalonia.Media.Brushes.Transparent
                 };
 
@@ -344,6 +346,7 @@ namespace PinayPalBackupManager.UI.UserControls
                     WindowStartupLocation = WindowStartupLocation.CenterOwner,
                     CanResize = true,
                     ShowInTaskbar = false,
+                    Topmost = true,
                     Background = Avalonia.Media.Brushes.Transparent
                 };
 
@@ -385,6 +388,7 @@ namespace PinayPalBackupManager.UI.UserControls
                     WindowStartupLocation = WindowStartupLocation.CenterOwner,
                     CanResize = false,
                     ShowInTaskbar = false,
+                    Topmost = true,
                     Background = Avalonia.Media.Brushes.Transparent
                 };
 
@@ -427,6 +431,7 @@ namespace PinayPalBackupManager.UI.UserControls
                     WindowStartupLocation = WindowStartupLocation.CenterOwner,
                     CanResize = false,
                     ShowInTaskbar = false,
+                    Topmost = true,
                     Background = Avalonia.Media.Brushes.Transparent
                 };